fix(database): env.AUTHORIZED_USERS is a string

+++ .scripts/generate_database_secrets.py
@@ -1,7 +1,8 @@
 import base64
-import json
 import os
 import subprocess
+
+import ujson
 
 current_dir = os.path.dirname(os.path.abspath(__file__))
 root_dir = os.path.dirname(current_dir)
@@ -37,19 +38,7 @@ for user in deployments:
 
 
 def deploy(data: list[dict[str, str]]):
-    fragments = []
-
-    def convert_dict(data: dict[str, str]) -> str:
-        fragments = []
-        for key, value in data.items():
-            fragments.append(f"{key} = {json.dumps(value)}")
-        # print(fragments)
-        return "{{ {} }}".format(", ".join(fragments))
-
-    for part in data:
-        fragments.append(convert_dict(part))
-
-    result = "[ {} ]".format(", ".join(fragments))
+    result = ujson.dumps(data)
     print("len(result) =", len(result))
     os.chdir(database_dir)
     subprocess.run(
@@ -59,6 +48,10 @@ def deploy(data: list[dict[str, str]]):
     os.chdir(current_dir)
 
 
-deploy(authorized_users)
-with open(output_path, "w", encoding="utf-8") as fout:
-    json.dump(authorized_users, fout, indent=2)
+if __name__ == "__main":
+    if os.path.exists(output_path):
+        raise SystemExit
+
+    deploy(authorized_users)
+    with open(output_path, "w", encoding="utf-8") as fout:
+        ujson.dump(authorized_users, fout, indent=2)
